Fix preview & submit button on application form
If the submit button is disabled before submit event is processed by the
borwser, the name/value on the button is not included in the form value.

This PR fixes that by using setTimeout.

There are other minor updates to js, e.g. iterating on submit buttons
instead of processing only the first submit button.

Some formatting update html is also there - helps keep html readiable
and for better git diff in future.