Fix typos in docs, code and comments and "typos" to pre-commit
Used pyproject.yml to file exclusions

This packages is quite fast and has very low false positives.

Given the number typos worth keeping it was able to find, I think it
will be quite helpful for it to be part of our pre-commit and lint
pipeline. It's quite easy to add new spellings or regex based ignore
rules.

https://github.com/crate-ci/typos