Fix typos in docs, code and comments and "typos" to pre-commit (#4512)
This package is quite fast and has very low false positives.

Given the number typos worth keeping it was able to find, I think it
will be quite helpful for it to be part of our pre-commit and lint
pipeline. It's quite easy to add new spellings or regex-based ignore
rules.

The exclusion and typo rules can be add to `pyproject.yml`. 


Read more at https://github.com/crate-ci/typos