Enable group stream boxes when in draft and edit mode (#4032)
Fixes #4031

They are disabled when editing a proposal in general, but in the case of
a draft, the users sometimes need to be able to edit them in order to
finish an application.

This fix may be incomplete, as it does not allow the group forms to have
the full functionality of hiding the followup forms when "No" is
selected, and showing them when "Yes" is selected.

Co-authored-by: Fredrik Jonsson <frjo@xdeb.org>