Fix page being blank on small screens due to table filters (#4525)
There seems to be an extra fixed DOM element present that activates only
on
small screens and takes over the whole screen; nothing is visible on the
page if there is a table rendered with the new table filter code.

This PR removes the extra style at the table filter top level wrapper
and also adds some gap between two filter items in the button form.
Here is the original image: