Async exporting of applications (#4471)
This PR makes the generation of submission exports async as (especially
in production) the generation of these files can take a long time and
cause a request timeout. This allows the user to do other tasks while
the file generates, and will notify them when the file is ready for
download via a new task on their dashboard or a green indicator on the
submission export button in the "All Submissions" view.

**Changes TLDR**:
- Added tasks to `hypha.apply.funds` module to handle the generation of
the submissions CSV
- Added a `SubmissionExportManager` model that keeps track of the status
of the generation & the finalized data
     - Each user can only have one active model/download
- Added a check in the nprogress event to allow for the disabling of the
progress bar while polling
- Preserved all sync functionality for users that don't need the
background worker, which is the default behavior

This is a WIP that should be completed early next week, but wanted to
get eyes on it early for any critiques. Steps left:
- [x] Add unit tests
- [x] Make a management command to remove `SubmissionExportManager`s
older than a day
- [x] Error handling for a failed generation
- [x] Triggering updating of  `My Tasks` when downloading from dashboard

**Running locally**
1. Get a Redis instance setup locally (I used docker, ie. `docker run -d
-p 6379:6379 redis`)
2. Install requirements
3. In a separate terminal from the running `make serve` (but with same
environment vars), run `celery -A hypha.celery worker -E
--loglevel=info` to start celery

## Demo

*Generating and downloading the exported CSV*


https://github.com/user-attachments/assets/71758bf1-2a0c-461f-9a4a-6942bf77b694

*Failed generation of a CSV*


https://github.com/user-attachments/assets/bb3029b2-0eb2-471a-930e-d8c85ff4ed6f