Better error handling for dates, added submitted filter criteria (#4549)
Fixes #4548. This handles an invalid date the same as
invalid/non-existent query params rather than erroring out. ~~In the
long term, it would be better to indicate to the user that something
wasn't right with their query though or no results were brought.~~ When
an invalid query is entered, it'll show the no results prompt!

I also added the submitted drop down filter, as staff have been using
that a bit recently