Fix submission filters queryset issue (#4572)
Fixes #4558 

All the code removed in this PR were used for previous/old filter
patterns. It was used to update filter datasets dynamically but now we
are treating it in a different way and handling those things in
submission all view so it is not needed here. It was also messing up
with filtered queryset for current implementation.
Now it should show/add/remove 0 submission if filtered query doesn't
belong to any submission.