Fix report file path issue and duplicate file issue, Fix file path issue for Project and SOW Form (#4580)
Fixes #4574 

As reports are project documents so we added a different path for its
files but we used ReportVersion instance checks and missed Report
instances so reports files and its version files were having different
path and were creating issues. We could solve it in that way as well but
I realise as projects are also the extension of submission and we have
been trying to merge them so keep everything similar with same url
pattern.

Now it uses submission id to generate path and render the file similar
to other submission files.

Now, it also handles the duplicate file issue on every upload. It was
resolved for the application form but not for the report forms.