Fixed incorrect pairing of revisions fields & revision CSS (#4608)
<!--
Thanks for contributing to Hypha!

Please ensure your contributions pass all necessary linting/testing and
that the appropriate documentation has been updated.
-->

<!--
Describe briefly what your pull request changes. If this is resolving an
issue, please specify below via "Fixes #<Github Issue ID>"
-->
Fixes #I4607. Some basic improvements to the revision view, ensures that
fields are always in order and makes the CSS more consistent through the
revisions. This is the first of a few enhancements to the revision view.


## Test Steps
<!-- 
If step does not require manual testing, skip/remove this section.

Give a brief overview of the steps required for a user/dev to test this
contribution. Important things to include:
 - Required user roles for where necessary (ie. "As a Staff Admin...")
- Clear & validatable expected results (ie. "Confirm the submit button
is now not clickable")
- Language that can be understood by non-technical testers if being
tested by users
-->

- [ ] Ensure fields in the application revision view are in order and
paired correctly with their header
- [ ] Ensure the "altered" yellow margin appears on all relevant
elements, not just `<p>`