Reverted to using difflab for revisions comparsions [redux] (#4610)
<!--
Thanks for contributing to Hypha!

Please ensure your contributions pass all necessary linting/testing and
that the appropriate documentation has been updated.
-->

(The other merge got botched so this is essentially a copy of #4609)

This is a temporary reversion to difflab for the revisions comparison as
the htmldiff library was using too much memory and timing out when the
compare view was loaded.

Eventually would be nice to make an html type diff work as lots is lost
in the stripping of the html tags.

## Test Steps
<!-- 
If step does not require manual testing, skip/remove this section.

Give a brief overview of the steps required for a user/dev to test this
contribution. Important things to include:
 - Required user roles for where necessary (ie. "As a Staff Admin...")
- Clear & validatable expected results (ie. "Confirm the submit button
is now not clickable")
- Language that can be understood by non-technical testers if being
tested by users
-->

- [ ] Ensure the compare view works as expected (side by side revisions)
and matches the themes vibe